requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.4
langgraph>=0.0.20
python-dotenv==1.1.1
ddgs==9.1.2
beautifulsoup4==4.13.4
googlesearch-python==1.3.0
streamlit==1.32.0
requests>=2.31.0
